Removed factories package, integrated into the prefabs classes as part of object creation
